server/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
chatterbot-corpus==1.2.0
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.2
Flask==2.3.2
Flask-Cors==4.0.0
frozenlist==1.3.3
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
MarkupSafe==2.1.3
mathparse==0.1.2
multidict==6.0.4
mysql-connector-python==8.0.33
nltk==3.8.1
openai==0.27.8
packaging==23.1
Pint==0.22
pluggy==1.2.0
protobuf==3.20.3
PyJWT==2.7.0
pymongo==3.13.0
pytest==7.4.0
python-dateutil==2.7.5
python-dotenv==1.0.0
PyYAML==3.13
regex==2023.6.3
requests==2.31.0
six==1.16.0
SQLAlchemy==1.2.19
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.3
Werkzeug==2.3.6
yarl==1.9.2
